requirements.txt
openai==0.28.1
gradio==4.19.2
pyttsx3==2.90
gTTS==2.5.1
python-dotenv==1.0.1 